requirements.txt
streamlit>=1.31.0
boto3>=1.34.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
